// src/redux/dashboard/types.ts
export enum DashboardActionTypes {
  FETCH_DASHBOARD_REQUEST = 'FETCH_DASHBOARD_REQUEST',
  FETCH_DASHBOARD_SUCCESS = 'FETCH_DASHBOARD_SUCCESS',
  FETCH_DASHBOARD_FAILURE = 'FETCH_DASHBOARD_FAILURE',

  FETCH_SPINDLE_STATUS_REQUEST = 'FETCH_SPINDLE_STATUS_REQUEST',
  FETCH_SPINDLE_STATUS_SUCCESS = 'FETCH_SPINDLE_STATUS_SUCCESS',
  FETCH_SPINDLE_STATUS_FAILURE = 'FETCH_SPINDLE_STATUS_FAILURE',

  FETCH_ERROR_WINDOWS_REQUEST = 'FETCH_ERROR_WINDOWS_REQUEST',
  FETCH_ERROR_WINDOWS_SUCCESS = 'FETCH_ERROR_WINDOWS_SUCCESS',
  FETCH_ERROR_WINDOWS_FAILURE = 'FETCH_ERROR_WINDOWS_FAILURE',

  FETCH_SENSOR_DETAIL_REQUEST = 'FETCH_SENSOR_DETAIL_REQUEST',
  FETCH_SENSOR_DETAIL_SUCCESS = 'FETCH_SENSOR_DETAIL_SUCCESS',
  FETCH_SENSOR_DETAIL_FAILURE = 'FETCH_SENSOR_DETAIL_FAILURE',
}